Attempted fixed joystick but decided to cut it for now.
Fixed joystick has an image component that anchors in the corner when the floating joystick has the image component but takes up the whole screen. This messes up our dash and moonball mechanic. So at this point i would need to rewrite the buttonIndicator script (our dash script) and fixed joystick script to work completely differently then the floating joystick.